pyproject.toml
[tool.poetry]
name = "drf-coursework"
version = "0.1.0"
description = ""
authors = ["Матвей Новиков <originsik@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flake8 = "^7.1.1"
black = "^24.10.0"
isort = "^5.13.2"
django = "^5.1.4"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.10"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.6.0"
drf-yasg = "^1.21.8"
redis = "^5.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
